+++ djangae/sandbox.py
@@ -260,3 +260,40 @@ def allow_mode_write():
     stubs.FakeFile.ALLOWED_MODES = frozenset(new_modes)
     yield
     stubs.FakeFile.ALLOWED_MODES = original_modes
+
+
+def allow_modules(func, *args):
+    """
+        dev_appserver likes to kill your imports with meta_path madness so you
+        use the internal ones instead of system ones, this wrapper reloads the
+        modules and patches the google internal ones with the __dict__ from the
+        system modules, this seems to be the cleanest way to do this even though
+        it is a bit hacky
+    """
+    def _wrapped(*args, **kwargs):
+        import sys
+
+        import subprocess
+        import os
+        import tempfile
+        import select
+        # Clear the meta_path so google does not screw our imports, make a copy
+        # of the old one
+        old_meta_path = sys.meta_path
+        sys.meta_path = []
+        patch_modules = [subprocess, os, tempfile, select]
+
+        for mod in patch_modules:
+            _system = reload(mod)
+            mod.__dict__.update(_system.__dict__)
+        try:
+            return func(*args, **kwargs)
+        finally:
+            # Restore the original path
+            sys.meta_path = old_meta_path
+            # Reload the original modules
+            for mod in patch_modules:
+                _system = reload(mod)
+                mod.__dict__.update(_system.__dict__)
+
+    return _wrapped
